Cleanup types for editor undo management

+++ src/txt/proto.h
@@ -212,9 +212,9 @@ COMMON(status)	undoBufferSizeTextBuffer(TextBuffer tb, Int size);
 COMMON(status)	markUndoTextBuffer(TextBuffer tb);
 COMMON(status)	resetUndoTextBuffer(TextBuffer tb);
 COMMON(status)	checkpointUndoTextBuffer(TextBuffer tb);
-COMMON(void)	register_insert_textbuffer(TextBuffer tb, long int where, long int len);
-COMMON(void)	register_delete_textbuffer(TextBuffer tb, long where, long len);
-COMMON(void)	register_change_textbuffer(TextBuffer tb, long int where, long int len);
+COMMON(void)	register_insert_textbuffer(TextBuffer tb, size_t where, size_t len);
+COMMON(void)	register_delete_textbuffer(TextBuffer tb, size_t where, size_t len);
+COMMON(void)	register_change_textbuffer(TextBuffer tb, size_t where, size_t len);
 
 /* /swi40/jan/src/pl/packages/xpce/src/txt/utf8.c */
 COMMON(char *)	F_UTF8_GET_CHAR(const char *in, int *chr);
